[build] Change how in_default_toolchain is computed.
Do not rely on the definition of toolchain_variant to determine
whether BUILDCONFIG.gn is evaluated in the context of the default
toolchain. Instead make the variable a build argument that
defaults to true, and that is forced to false by any toolchain()
definition in the build.

Change-Id: Iefc6926b04f0934a4225ca2e346271a7ec41db0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1389833
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>