[kernel][pdev] Define a "clocks and pmic" pdev interface.
Define a pdev interface to abstract clock and pmic manipulation which
may need to be performed by the kernel immediately before/after
entering/leaving certain low power suspend states.

Currently, these hooks are needed in order to shut of clocks needed to
hit VDD-min after preparing the debug UART for suspend on Sorrel.  It
is not clear what the final design of this is going to end up looking
like: the topic turns out to be rather complicated.

In the meantime, however, we need to make sure the clocks are off, or
we cannot hit the min power state.  So, for now, this is a placeholder
for a temporary design meant to unblock others, and may change in the
future.

Bug: 417556966
Change-Id: I93dd8d770bb542e763ebb2f8259811e61fbd5649
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386967
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>