[ffx] in direct mode don't wrap all errors in User
In direct mode, we were wrapping every error in User, because we want
to make sure to attach any connection errors. But if the error didn't
start as a User error, this would cause problems (e.g. for programs
that exited with Error::ExitWithCode). Instead, just wrap errors
when we already started with User or Unexpected.

Change-Id: I6a6a69647076f92da33b0565cb7ac065d728a0da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1389193
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>