[kernel][clocks] Implement a moonflower clocks+pmic driver.
Implement a placeholder clocks+PMIC driver for moonflower.  Currently,
this driver will handle shutting down the clock tree feeding the debug
UART after the UART has been prepared for suspend, and before the
system actually attempts to enter suspend.

In the near future this will be extended to also request that the RPM
shut down the voltage rail feeing the PLL at the root of this tree,
which should allow the system to be able to make it all of the way
into VddMin.

Bug: 417556966
Change-Id: Iffe71e53c7ced9080a762272b4c8e42fc8fd486f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1387885
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: John Grossman <johngro@google.com>