[debugger] Cache currently active non-elf components
Previously, the start and stop events for these components were fired up
to the DebugAgent, but there was no way to query them after that. That
means that if a filter which was intended to match one of these
components wasn't installed before the component was started then the
filter would never match.

This change adds the caching mechanism, but doesn't make use of the
cached values yet, that's done in a follow up.

Bug: 373824677
Change-Id: I6a6a696497d6d57c999b032a71ea84949dc32215
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1391013
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>