[assembly] ffx product bisect
This adds the "bisect" plugin to the ffx tool, which can be used to
quickly find which version of the assembly artifacts introduced a bug in
a given product bundle.

This CL implements a "longest dimension" bisection search strategy which
bisects the artifact list with the longest version list first, but it
provides an easy way to change the algorithm from the command-line.

Run using:
  ffx product bisect

Bug: b/441912787
Change-Id: Iedb8d2b37afd7dd4fd4bfa50e03c91fb892ef5c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1366464
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>
Fuchsia-Auto-Submit: Jerry Belton <jcecil@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>