[sestarnix] Keep security state alive in FromTask
When a FsNode is labeled using `FromTask`, we would hold a weak
reference to the Task. This means that once dead, we lose the label
and have to revert to unlabeled. Instead, make the TaskState shared so
that we can hold a strong reference to the latest state.

Bug: 450515764
Change-Id: I41a9e5c13f40769eda4811bb4977f7007e8c5abe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1388653
Commit-Queue: Ambre Williams <ambre@google.com>
Reviewed-by: Wez <wez@google.com>