[netlink][tests] Deflake integration test
`add_route_in_table_and_await_installed` assumes the next netlink
message will be the one we request to be installed. This is not a
sound assumption because a netstack managed route could race with
the user requested route. Change the helper to filter for the exact
route to address the race.

Multiply: netlink-integration-tests
Bug: 451645716
Change-Id: I6a6a696467792f0a7fac1148f64d3334017ce7d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1391593
Commit-Queue: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Sergey Ulanov <sergeyu@google.com>