[msd-vsi-vip] Submit multiple buffers per reset
We currently think a hang in this test is related to the reset
behavior. Submit multiple buffers per reset to check that it would
always be the first buffer that is causing the hang.

The logic here gets a bit more complicated because we can't
keep remapping a buffer at the same address. We create and map
the buffer once and submit it multiple times.

Bug: 439669949
Change-Id: Ifafe6de1cd585d1834656415806d6693f2e8851d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1391553
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>