[starnix] Move security dir under new core subdir.
The visibility modifiers in this module don't seem to work with
the re-exports I've set up to smooth out the transition, so this change
just makes the visibility path gross temporaraily. We'll be forced to undo
that part of this patch when the migration is finishing.

Change-Id: I704bcbbfed96db7d44b5746c46cee127526620d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1389207
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>
Reviewed-by: Wez <wez@google.com>