[build] Compute fuchsia and host toolchain tags in BUILDCONFIG.gn
Compute fuchsia_toolchain_tags and host_toolchain_tags to reflect
the variant tags applied to Fuchsia and host artifacts by default,
based on the value of select_variant in args.gn.

Due to the way toolchain variant redirection is implemented, the
value of toolchain_variant.tags in the default toolchain is always
an empty list.

A flag like is_asan is true to indicate that Fuchsia binaries are
generated with the "asan" or "hwasan" variant. This value is the
same in all toolchain contexts.

On the other hand, a flag like is_ubsan is true to indicate that
the *current* toolchain implements undefined sanitizer, and will
always be false in the default toolchain's context, even if
select_variant has 'ubsan' or 'asan_ubsan' in it.

This CL does not change the values of these flags but:

- It adds is_fuchsia_asan and is_host_asan to make it clear that
  these flags relate to Fuchsia and host binaries, and not the
  current toolchain.

  is_asan is now defined as an alias to is_fuchsia_asan to
  soft migrate definition in other build rules later.

- It adds is_toolchain_ubsan to make it clear that the flag
  relate to the current toolchain, and makes is_ubsan an
  alias for it, for similar reasons.

- It adds the "coverage_clang" and "coverage_rust" tags to
  differentiate between Clang and Rust coverage backends,
  since both compilers are currently using different LLVM
  backends (until https://fxbug.dev/42060061) is fixed.

  This is required to implement is_fuchsia_coverage_clang
  properly.

- It adds is_fuchsia_coverage_clang, and make is_coverage_clang
  and alias for it as a //third_party/ build rule uses it.

- Remove unused is_rust_coverage, is_cts_coverage variables.

This CL doesn't change the value of the flags though.

fuchsia_toolchain_flags and host_toolchain_flags will be used
in a future CL that fixes b/417952817

Bug: 417952817
Bug: 42060061
Change-Id: Iced3e33d0daa6e0390117dce1207b2d520f73aa7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1389835
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: David Turner <digit@google.com>