[sestarnix] Add `skip`&`expect_failures` for audit
Added the possibility to mark a test for skip action or for failure
expectation for audit expectations checking.

When a test is marked for skipping, the audit logs produced by it will
be drained and the checks will be skipped.

When a test is marked for failure expectation, the audit logs are
checked and errors are printed, but gTest failures are not added.

These tests should be added in the `audit_exceptions.json` file, which
will be merged and minified before being used by the checker.

Bug: 448860115
Change-Id: I2021ed8536a4da16e1654d2c0d35b5cf555453f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1382362
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Rareş Constantin <rarescon@google.com>