[fxfs][zxio] Zero out fsverity root hash in zxio not fxfs.
Zxio should not rely on the length of the returned value to be 64 bytes,
the returned value is a vector and can be variable length for the
different hash lengths. Zxio should be defensive as a client library.

Fxfs should also not extend shorter hashes out with zeroes. Just return
the hash as is. The caller will receive the length information from fidl
and should use it accordingly. Most of the time the caller probably just
wants to compare the hash values and doesn't want to have to inspect the
algorithm type to know whether they need to truncate the value or not.

Bug: b/399727919
Change-Id: Idd78d06738567ea223be20d3de816fcc364702b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386881
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>