[sestarnix] Add audit expectations update option
Added an option for the `run_on_linux.py` script to update the audit
expectations file. The script scans the output of the tests and parses
the JSON objects found, adding them to the expectations file at the end
of the script.

This option can be used with the filter option as well, the script
updates only the tests it finds in the output without modifying the
rest.

Notice - the update option is not 100% reliable, as the Linux kernel
syslog may corrupt the debug print of the audit JSON objects, therefore
a sleep of 1 second is added after the test suite finishes.

The permissive denial logs are skipped on Starnix, as they are not
logged as denials.

Bug: 448860115
Change-Id: I361e5cc4e13664262decfaa9e35360b8fc3d48c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1382365
Commit-Queue: Rareş Constantin <rarescon@google.com>
Reviewed-by: Wez <wez@google.com>