[sestarnix] Always check userspace tests audits
The tests are always checked and errors are generated if there are audit
expectations, but the tests do not generate them.

If the test is expected to fail, the audit logs will be checked and if
all logs match, it is added as a failure.

Bug: 448860115
Change-Id: Iafa5fd01197b8809562850de39beb1141168f9da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386644
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Rareş Constantin <rarescon@google.com>