[analytics] Add probes for new internal users
This a temporary change that will be reverted once the experiment is
done. The code were locally tested but no unit tests were added for the
temporary change.

The goal of the experiment is to see whether making enhanced analytics
the default for internal users will lead to many bot users being
recorded in analytics.

Bug: b/422225134

Change-Id: I6a6a6964f7abe6bf077595335580cbe18f51f330
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1387961
Commit-Queue: Rundong Du <rundongdu@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>