[f2fs_reader] Support verity for default encryption.
Default encrypted files will now migrate their verity settings for
default encrypted files. Since these files are being copied, the verity
data is not actually copied or mapped in, but rather the settings
including the salt are read out and then fxfs applies verity normally
using the same set of options.

The resulting hashes are verified to be the same now in the testing, but
the allocated size calculation is made more complicated to account for
the fact that the files will take up different amounts based on the
different formats for the merkle tree data and fsverity descriptor.

Bug: b/399727919
Change-Id: Ieb94cf3fc76bc342093c3fad7522aab16873b4e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386882
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>