[netstack3] Fix invariant on TCP BINDTODEVICE
A connected socket may _not_ be in the demux if the state machine has
moved to the closed or timewait state. In closed state we always remove
the socket from the demux, and in timewait the socket _may_ be removed
as part of reusing the slot (because of timewait reuse).

Encode those invariants in `set_device_conn` and add a regression test
for the panic we were hitting (found by syzkaller).

Fixed: 388656903
Change-Id: Id5a1456c25d907775ca6afaa853bce84701a27e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1391634
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>