[debugger] Overhaul Filter identification
This change accomplishes two primary goals:

  1. Disambiguate filter creators while maintaining backwards
     compatibility.
  2. Add an event that tells clients about backend created filters
     (e.g. the filter that gets installed from a recursive filter).

Previously, the backend would silently create filters that would get
wiped upon any updates to filters by clients. There was also a segmented
approach to how clients (namely zxdb and the DebugAgent FIDL server)
identify their own filters. Now, this is encoded into the debug_ipc
protocol, with additional room for expansion. The changes to
debug_ipc::Filter and related objects are ABI compatible with the
previous implementation (the Filter::Identifier object is sent as an
opaque uint32_t across the wire, the same as before). We have also not
sacrificed the feed-forward client-allocated ID approach that existed
before.

In particular, the most significant byte of the 32 bit value has now
been segmented bitwise into 8 distinct values. Three of these bits are
allocated to the existing three entities in the system today - zxdb,
DebugAgent, and the FIDL server. As was the case before this change, the
FIDL server maintains the highest bit for its own filters, with zxdb and
DebugAgent receiving their own allocations. All clients now use truly
opaque random values in the low 24 bits of these values to discourage
forming additional dependencies on the filter identification values.

These changes are a necessary prerequisite to landing support for
recursive job-only filters, which are intended to be used in test
timeout situations in infrastructure.

[g-review]

Bug: 303314982
Bug: 373824677
Change-Id: I6a6a6964b6c9fda6272f14b4a7ee0fd83630168e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1391014
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>