[debugger] Support recursive job-only filters
These have a particularly distinct use case of giving us the ability to
easily be able to attach to all processes spawned in a particular
component realm. Specifically, this is extremely useful for the FIDL
interface where other entities such as TestManager know a moniker suffix
of all the components that they will spawn, but don't necessarily know
all of the monikers that a particular test will spawn within its realm.

In this situation, TestManager can simply install a recursive job-only
filter so that DebugAgent is aware of all process events in the realm,
and in the event of a timeout, can vend backtraces back to TestManager
for all components in the entire realm, which is useful for tests that
make use of things like RealmBuilder and may have components that are
not the primary test process get stuck. In other cases where a process
crashes, we can operate as normal and return the singular backtrace for
the crashing thread.

[g-review]

Bug: 303314982
Bug: 373824677
Change-Id: I6a6a6964d3f6ef6151f54198d6dc7d95932a4202
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1391015
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>