[kernel][phys] Leave a placeholder for a virtual heap starting address
physboot will be in the limited business of allocating a virtual address
range for the kernel's virtual heap. Right now we construct with kernel
address space with hand-off data below the kernel load image and
special, 'first-class' mappings above it. It seems most reasonable to
place the virtual heap reservation after the first-class mappings
(though no mappings for it are to be made in physboot). Since there is
no configured upper boundary for the first-class mappings, we add logic
to their allocator to mark it as done and return its ending boundary,
which can later feed into the decision on where to place the virtual
heap range.

Bug: 446675650
Change-Id: Ib7e2af82f1f88dfcce0bdee9427e54e847c70950
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374249
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>