[audio_recording] Support edge-triggered audio capture
This CL implements a new method for capturing audio using the
audio_recording component: Edge-triggered audio capture.

Clients may queue a future capture by specifying parameters and then
asynchronously wait for the result of this operation. This is an
improvement on existing test scenarios in which we simply record a long
duration of audio and hope that it includes the part we care about.

Not only is this method more reliable, it is faster since we can eagerly
stop capturing when no more audio is observed for some duration.

Fixed: 433363761

Change-Id: Ib91b04ec86fb6529ead3f5988acc1af9dd756a7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364820
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Henry-Michel Cantave <hmcantave@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>