[bazel] Add cc_source_library_zx() macro
The macro is a thin wrapper around `cc_library()` that bazel2gn
converts to a `zx_library()` with `sdk = "source"`.

Rename `zxAttrMap` to `zxInIDKAttrMap` to reflect that it does not apply
to non-IDK libraries.

NO_IFTTT=The change to zx_library.gni is specific to GN. The Bazel macro has a default value.

Bug: 429377203, 429374432
Test: fx test --host bazel2gn_tests
Change-Id: I13949d17312b5e36aa79054d0f1491f4ea740b53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1392905
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: David Turner <digit@google.com>