[memory_sampler] Enable profiling non-core components
Now that fuchsia.memory.sampler.Sampler is part of the SDK,
non-platform components can start opting-in to the instrumentation
necessary for profiling. They typically live under the
product-specific session, so for the capability to be useful, it must
be forwarded from core to the session, which this change does.

Bug: 412699984

Change-Id: I6f32a60d9e9ce912f3bb97d307e07f88b5fcf384
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1388593
Reviewed-by: Akira Baruah <akirabaruah@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>
Fuchsia-Auto-Submit: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: Étienne J. Membrives <etiennej@google.com>