[test-pilot] add use_if_defined property
This CL adds a use_if_defined property to output processor entries. The
property causes test-pilot to suppress the invocation of an output
processor if one of a list of test parameters is not defined. The
primary use case for this is to optionally generate additional output
based on whether an destination path for that output is provided.
Botanist will use this to generate an output summary in result_db
format.

Bug: 443323806

Test: new and update unit tests
Change-Id: I27041555113a28f47dc23a3b6e41436f4fa66104
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1387888
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>