[starnix][vsock] Perform internal operations as kernel
As an implementation details, connected vsock sockets hold a
FileHandle to an pipe to the Fuchsia world. This is created in a
"kernel" context, which means the file descriptor is labeled as
"kernel", triggering fd use checks. Avoid these checks by switching to
kernel credentials instead.

Bug: 442899275
Change-Id: Ie0c0a017c2bd92a62eb0742d73fb8e65014f0a80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1392293
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Ambre Williams <ambre@google.com>