[kernel][vm] PeekIsolate calls SynchronizeWithAging once
The intent behind synchronizing with aging is that we want to have the
best chance to be able to find reclaimable pages. If there was an aging
outstanding at the time of the call, it's possible that we falsely claim
we have no reclaimable candidates. The synchronization ensures that we
have the most accurate mru_gen_ *before* we compute the LRU target that
the lru_gen_ can be advanced to before returning.

However, the synchronization should happen only once, and not on every
iteration of the loop. If we update the mru_gen_ more than once, it is
possible to get into a state at OOM where we've burnt through all of the
non-active page queues, and the mru_gen_ and lru_gen_ increment in lock
step, without finding any reclaimable candidates, never terminating the
loop.

Bug: 423900604
Run-All-Tests: true
Change-Id: I02df2e70e253838a8f6122c720eb0c84ca93284a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1394213
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>