[usb_rs] fix file descriptor leak
When dropping the Interface, the inner's reaper_thread needs to be
unparked.

The reaper_thread has a Weak reference to InterfaceInner, which has a
reference the the FileDescriptor for the USB Device. The reaper_thread
is also parked if there are no URBs to reap. Since it is parked while
dropped, and it has the Weak<InterfaceInner> still, the InterfaceInner
will never be dropped.

By unparking the thread, it will `continue`, attempt to upgrade the
inner, get None, exit the loop, then the thread, allowing the
InterfaceInner to be dropped along with the open File.

Test: before this change run ffx target discover -f -t 4 and inspect the procfs
entry for the process. Notice that it keeps the file descriptors for the USB
devices still open (fd directory). Apply this change and re-run the command
and notice how the file descriptors are closed.
Bug: b/450617502
Fixed: b/450617502
Change-Id: Ib0ca9a51a932c0ba305987aa898404216a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1395617
Commit-Queue: Colin Nelson <colnnelson@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>