[sestarnix] Use a Mutex in audit subsystem
Switched to a single Mutex over the audit sink reference and its queue
because atomic operations cannot ensure there will be no races between
multiple threads that manipulate the state of this reference.

If performance will be a concern in this subsystem, RCU may be used, but
the audit queue should become an internal part of the
`AuditNetlinkClient` and a better idea would be to use a kernel thread
to ensure the logging threads do not have to check the registered audit
sink while adding messages.

Linux uses a locking mechanism with RCU while updating the reference of
the audit sink.

Added a netlink audit userspace test which checks what happens if an
unregister request is sent while there is no audit client registered.

Bug: 447104596
Change-Id: I049617bfd9b4dbf40a659fb67007117667bee5be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1393653
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Rareş Constantin <rarescon@google.com>