[power][lib] Remove optimization
Previously after acquiring a wake lease we would update the
internal state of the wrapper class to consider the system
resumed. This is under the assumption that even though the object
had not observed the resume callback, it would soon.

This optimization rests on a false assumption, is unlikely to be
useful, and would create a very difficult debugging task if we
ever hit the bad edge.

This is a potentially bad assumption depending on the internal
implementation of SAG. SAG is only guaranteeing the system stays
awake while the wake lease exists, not that we go through resume
callback. If the acquired wake lease comes and goes quickly
enough, it may be possible resume callbacks never happen. If we
ever hit this edge it would likely be extremely difficult to
identify as the culprit.

Additionally, this optimization is likely not all that useful. If
a client called `ManualWakeLease::Start` twice in a row we still
take a fast path because a lease is held internally. The
optimization is only useful if a client calls
`Start`->`End`->`Start` such that the second `Start` call is
before the resume callback.

Change-Id: Icafc92aab5d747c75502f74b86a0031c9aeafdfb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1396536
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Justin Mattson <jmatt@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>