[ffx] check "ffx -c shared_data=<dir>" in strict
When in strict mode, if a config variable includes the $SHARED_DIR
mapping, we verify that ffx was invoked with "-c shared_data=<dir>".
If not, we return an error when retrieving the variable.

Note that in this change, we rely on the fact that there is
never a use-case where we need to support "$SHARED_DATA" being
mapped _multiple times_ in a single config value. Therefore we
can use the new `try_replace()` function which only maps
the first occurrence in a given variable.

Test: Added E2E test to verify error on missing shared_data in strict
mode.

Fixed: 384787070
[g-review]

Change-Id: I6a6a69645ba7443748b4e80021c2b5fcc09b6fc0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1394258
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>