[kernel][crashlog] Fix the uptime estimate.
A while ago, the concept of "boot" time was introduced along side of
"monotonic" time.  They are basically the same, but "monotonic" time
was defined to stop when the system was suspended, while "boot" time
keeps running.

After this happened, there were a number of places where uses of mono
needed to be converted to uses of boot when it mattered.

The RAM mappable crashlog was one of these places.  It was properly
updated to record both the boot and monotonic times in its header, so
that the estimates could be recovered in the case of a totally
spontaneous reboot (think, HW WDT, overtemp, brownout, etc)

This is a place that got missed.  The UPTIME estimate which is
recorded during a non-spontaneous reboot was being recorded using mono
time when it should have been boot.  Fix this.

Change-Id: Ia4e93b3032889be145760883d5a3c1bf53fe59c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1396733
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>