[kernel][timer] Strengthen sync guarantees.
There are a couple of variables (`cancel_` and `active_cpu_`) in Timer
state which responsible for synchronizing some operations from outside
of the global timer lock.  Specifically, when canceling a timer, and
losing the cancel race.

1) When the timer fires, TimerTick is holding the lock and removes
   the timer from the pending timers list, and sets active_cpu_ to
   indicate that there is a callback in flight.  Then, it drops the
   lock and starts the callback.
2) Cancel acquires the lock and discovers that the callback is already
   in flight.
3) Cancel drops the lock, and spins on active_cpu_ waiting for the
   callback to be finished before returning control to the caller.
4) TimerTick re-acquires the lock, and marks the callback finished via
   active_cpu_
5) Cancel now sees that the callback is finished, and returns false to
   its caller to indicate "You lost the cancel-race.  The timer went
   off first, and its callback is now complete."

Prior to this CL, these variables were atomics being accessed using
relaxed semantic.  There is a theoretical problem with this.  In step
5, we told our caller "The callback is now complete" via our return
code.  This is true.  However for most users here, the implication
(and very reasonable expectation) is "The callback is complete, and
the writes performed by the callback are now visible to you".

Relaxed semantics are not going to give us this extra guarantee, which
is a problem.  All they tell is is that TimerTicks has finished, not
what it did in the callback.

Fix this, but do so in a slightly fancy way.  Most readers, and *all*
writers, of these variable do so inside of the lock.  If this were
true everywhere, we would not even need to be using atomics.  Readers
who are loading the variable from outside of the lock need to be using
acquire semantics, while writers need to be using release semantics to
establish a syncs-with relation with the outside-of-the-lock readers.

But readers are inside of the lock, and don't really need acquire
semantics.  All of the writers are inside of the lock, and the lock
provides us all of the synchronization we need.  Relaxed is good
enough here.

So, introduce a small wrapper class for a plain old data type T.  The
wrapper class provides 3 methods to access the underlying atomic T
storage.

1) store_locked.  Requires the lock, stores with release.
2) load_locked.  Requires the lock, loads with relaxed.
3) load_unlocked.  Excludes the lock, loads with acquire.

+ Ops 1 and 2 sync using the lock, which is statically required.
+ Ops 1 and 3 sync using acq/rel.
+ Readers cannot accidentally select the wrong version of load.  If
  they try to call the locked version without holding the lock, the
  analyzer will catch them, and the same is true in the other
  direction.

Finally, convert the variables to use the wrapper and update the call
sites.

One last note.  This potential issue was discovered when pondering
potential causes of bug 453005631, and is being addressed
pro-actively.  That said, we do not believe that the old relaxed-only
semantics could have been the cause of 453005631, so the investigation
continues.

Bug: 453005631
Change-Id: Id3bef7521cff9a071d31109e5b1dbfb8614bb9c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1396813
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>