[sysmem] new required_max_size_list
The required_max_size constraint field is deprecated in favor of
required_max_size_list which avoids the downsides of having a single
width and height for this, which in some situations doesn't make
sense.

A main downside of the old required_max_size field is there's no
way to say "I want both landscape and portrait images to fit in the
buffer(s), but I don't need a single image with the biggest width and
height across landscape and portrait to fit in the buffer(s)." With
required_max_size_list, this can be specified with two entries in
required_max_size_list with the relevant max landscape and max
portrait dimensions.

Change-Id: I2cafdcfa5d2f85fdbbc2d76e98ea76bd6b1adaea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1388357
Reviewed-by: Emircan Uysaler <emircan@google.com>
API-Review: Emircan Uysaler <emircan@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>