[netstack3] Collect matcher TryFromFidl impls
We're going to be adding more for socket diagnostics, and all of these
are widely reusable, so it's worth having them all condensed in a single
place.

As part of this, we also pulled implementation out of filtering and
routes for InterfaceMatcher and MarkMatcher.

Bug: 449158018
Change-Id: I6a6a6964995590af282c2789a3881a7f8b0d9673
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1392637
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>