[fidl][rust] Improve response ergonomics
This change adds the `Respond` and `RespondErr` traits to allow the
bindings to generate more ergonomic respond methods. To do so, it also
relaxes some of the constraints around the usage of `EncodeRef` to allow
dynamically-sized types to be encoded, as well as
references-of-references.

Fixed: 441822506
Change-Id: I28ed9819e0dc6b7f7947c1db6851e35e57795189
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1389314
Reviewed-by: Megan Batty <mgnb@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>
Owners-Override: James Robinson <jamesr@google.com>