[starnix][utc] Simplify the computation of the initial UTC clock transform
The old computation was based on a straightforward change from the
monotonic timeline to boot timeline. Mono timeline needed to compute
the offset because the mono timeline didn't start from zero on boot.

Since the boot timeline is defined as starting from zero on boot,
we can simply insert a reference of `(Boot(0), Utc(backstop))`
and be done with it.

This change seems small and inconsequential. But, it makes it much
easier to migrate to memory mapped clocks, since it will become
obvious that this value is constant, as it depends only on values
that are known at compile time.

Bug: 422533641
Change-Id: I3fa85b11d9f422055d00ce8f7894635f71526bb8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1398674
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: James Robinson <jamesr@google.com>