[testrunner] Expect TestResult as test output summary schema.
Originally we expected a list of TestCaseResult, but ffx test/test-pilot
returns a summary with its own outputfiles which we should use if
provided as well.
Also update the go_test_parser to provide the stdout from a
test case as the fail reason if it fails and use the
runtests.TestDetails throughout the testrunner code instead of the
separate testrunner.TestResult.

Bug: 435723394
Change-Id: If120b8589f9d46381fb9c0634731e1a28578ed2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1379981
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>