[starnix] Remove stale GN args.
The 32bit arg in particular complicates splitting syscalls out to
separate libraries and now that 32-bit support has been stabilized
I don't think we need to be able to turn it off with a GN arg
anymore.

Based on an extremely unscientific poll in chat, none of the other
args are currently in use.

This change also fixes up some test expectations around unified
aspaces that had gotten a bit out of sync. It seems that we now
pass the partial I/O tests when the file is backed by fxfs, but
not when it's backed by Starnix's tmpfs.

Change-Id: I2845f27d5984c24315c48e66197f7788b834c97a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1389301
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>