[reachability] Improve gateway calculation
Previously, when any gateway was discovered, both protocols had their
LinkState set to at least Gateway. This is not necessarily correct since
there may have been a v4 route to a gateway, which should not imply v6
Gateway connectivity. Improve this calculation to check for the protocol
of the routes to the gateways.

Fixed: 362762533
Change-Id: Ida6ea45d1e60d63730c1e5bb0d4bd4a193c8abf9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1369776
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>