[usb_vsock] Pause packets correctly set length
The pause packets were not setting the payload length, and the way that
write_unchecked worked made that overwrite part of the packet header.
This fixes it so the length is set and also prevents that particular
clobbering bug from recurring.

Change-Id: Iaa84b35c02d91e3eee31c2bc54f96f3396f9db0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1399396
Commit-Queue: Megan Batty <mgnb@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>